requirements.txt
Flask-Migrate==2.0.0
gunicorn==19.6.0
msgpack-python==0.4.8
Flask-Script==2.0.5
Flask-SQLAlchemy==2.1
blinker==1.4
Flask-Assets==0.12
jsmin==2.2.1
cssmin==0.2.0
Flask-Markdown==0.3
dropbox==7.1.1
Flask-Compress==1.4.0
Flask-HTMLmin==1.2
celery==4.0.2